week5/requirements.txt
openai
python_dotenv
pillow